spotify_project/requirements.txt
fastapi==0.110.1
pydantic==1.10.2
spotipy==2.22.0
starlette==0.37.2
itsdangerous==2.2.0
uvicorn==0.20.0
python-dotenv==0.20.0